chore: get jest to work again
