+++ package.json
@@ -7,6 +7,7 @@
         "start": "vite --port 3000",
         "build": "tsc -b && vite build",
         "preview": "vite preview",
+        "test": "jest test",
         "format": "yarn d2-style apply",
         "_lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
         "lint": "yarn d2-style check",
@@ -27,6 +28,8 @@
         "react-router-dom": "^6.24.1"
     },
     "devDependencies": {
+        "@babel/preset-env": "^7.24.8",
+        "@babel/preset-react": "^7.24.7",
         "@dhis2/app-runtime": "^3",
         "@dhis2/cli-style": "^10.5.1",
         "@ls-lint/ls-lint": "^2.2.2",
@@ -47,11 +50,11 @@
         "eslint": "^8.57.0",
         "eslint-plugin-react-hooks": "^4.6.2",
         "eslint-plugin-react-refresh": "^0.4.7",
+        "jest": "^29.7.0",
         "openapi-typescript-codegen": "^0.29.0",
-        "playwright-core": "1.44.1",
         "playwright": "1.44.1",
+        "playwright-core": "1.44.1",
         "styled-jsx": "^5.1.6",
-        "ts-jest": "^29.2.2",
         "typescript": "^5.2.2",
         "vite": "^5.3.1"
     }
